Don't block login on agent-tui
By setting agent-interactive-service RequiredBy sshd and systemd-logind,
we prevent the user from ever logging in to the host if the agent-tui
process fails. This makes it impossible to debug the failure.

We want to start the agent-tui whenever tty1 exists (which in practice
seems to be always, even if there is only a serial console). To ensure
that we don't block subsequent login to the console after a failure, use
WantedBy instead of RequiredBy. The other install dependencies are not
needed.