OCPBUGS-19444: Use changes to AgentClusterInstall during loading
In the Load() method of AgentClusterInstall, we sometimes make changes
to the input data to maintain backwards compatibility and such. However,
these changes had no effect because we always used the original file
read from disk as the generated output, even when it didn't match the
actual config.

Regenerate the data to use in the Agent ISO at the conclusion of
loading.