OCPBUGS-19092: Enable serial console for external OCI platform
When creating an Agent ISO for OCI, add the kernel argument
console=ttyS0 to the ISO/PXE kargs.

CoreOS does not include a console arg by default when using metal
as the platform because different hardware has different consoles
and specifying one can cause booting to fail on some,
but it does on many cloud platforms.

Since we know when the user is definitely using OCI
and we know the correct settings for OCI,
we should set them up automatically.

Signed-off-by: Pawan Pinjarkar <ppinjark@redhat.com>