OCPBUGS-19093: Do not start agent-tui if no graphical console available
If there is only a serial console, do not start agent-tui since it will
not be visible.

We assume that there will be a graphical console if there is a
framebuffer device (/dev/fb0). On x86 an alternative would be to check
for the existence of /sys/class/vtconsole/vtcon1. However, in the
absence of a range of hardware to test, the former seems more likely to
work on other platforms as well.