OCPBUGS-19688: Enable agent-tui on serial console
If there is only a serial console available on the machine, enable
agent-tui on that console since there is no graphical console.

Since only one serial console can be enabled at a time, it is safe to
use a template here to handle the fact that we don't know which serial
port the console will be on.