AGENT-713: Use baremetal hosts from install-config if not defined in agent-config
Initial pass of using the baremetal hosts defined in install-config.yaml for the
agent installer host definition if not defined in agent-config.yaml. This
provides compatibility with the baremetal IPI configuration (more changes will
be coming for the BMC fields as described in https://issues.redhat.com/browse/AGENT-677).

Removed all of the related warnings except for BootMode since this field is not used and
we should let the user know if set.