pkg/infrastructure: generic stages interface
Introduces an infrastructure package with a stages interface to
allow various infrastructure providers. The stages interface is
based on the existing pkg/terraform/stages interface, but allows
each stage to have its own Provision (and Destroy) function which
can call any infrastructure provider in addition to Terraform.

Currently, the input to the provider is still expecting TFVars,
which the underlying stage provision logic can translate.