OCPBUGS-18455: Unable to disable external CCM for platform external
Adds a validation that cloudControllerManager field must be set to External
when using oci as the external platform.
The default value for cloudControllerManager is an empty string.
Pass cloudControllerManager field value through to the install-config overrides
so that assisted-service uses it in the install-config
which ultimately generates the Infrastructure manifest