research-ci/azure: isolate/mount etcd to dedicated data disk
This PR forces to creation Data Disk (additional disk) on the Control Plane, creating MachineConfig to format and mount the etcd path (`/var/lib/etcd`) to the new device.

This is an experimental setup to be used in CI to target stability, resilience, increase performance, and reduce costs on Azure deployments.

The PR also forces the Control Plane VM Type to 5th generation (v5), decreasing the size to half, as the preliminary numbers point to keeping the same performance.

The plan is to run A/B periods in an isolated branch, measuring it within one week of execution.

Preliminary results: https://github.com/openshift/release/pull/44367