GCP Custom DNS work
Added new stage to terraform to extract Load Balancer information.

Added new step to the installation to grab the output from the stage above.

Added the lb config map. The data shows up on the bootstrap node, but it is incomplete.

Working on a way to add the correct data to the ignition file.

fixup