PowerVS: MULTIARCH-3791 Remove cloud connection reuse functionality and vars
One thing we can remove that will make removing cloud connections entirely later on easier is the ability to reuse a cloud connection.

This will require removing any vars in the installer that are specific to cloud connections and then removing the terraform specific to reusing a cloud connection.

This will leave us with an installer that can still be used in CI but is less entangled with cloud connections.

https://issues.redhat.com/browse/MULTIARCH-3791