Add yq-v4 to the upi-installer image for CI and copy yq3 from a previous stage's manifest-list image
While new steps are being implemented in Prow, users may want to pivot to yq-v4. This commit installs yq-v4 in the upi-installer image.
Moreover, by copying the yq-v3 and yq-v4 binaries from previous stages based on manifest-list images, we guarantee the yq binaries are the ones for the architecture of the upi image being built.