OCPBUGS-20364: azure: validation: validate defaultMachinePlatform
If only `defaultMachinePlatform` is specified, no instance type
validation was being done, resulting on installs with invalid instance
types when no `controlPlane` and `compute` sections are defined.

Let's make sure it's validated in case it hasn't been used by neither
controlPlane nor compute machines.

The only caveat is that `architecture` can only be specified in the
`controlPlane` or `compute` stanzas, so we're forced to assume `amd64`
when those stanzas are not present.