Enable platform validation for agent-based installs
Agent-based installation allows credentials to be optional.
Platform values are not validated if credentials are not provided.

Folder is now a mandatory field for agent-based installations. It
is optional for IPI and UPI. It is being made mandatory because
if credentials are entered and folder is left unspecified,
assisted-service will substitute in placeholder values for folder.
The placeholder value is "/datacenterplaceholder/vm/folderplaceholder".
When assisted-service generates the install-config, validation
will fail because "datacenterplaceholder" will not match the actual
datacenter that the user entered. Better to make it mandatory
and validate folder is entered when the agent ISO is being created.
Otherwise the user will not realize there is an issue until they
boot up the agent ISO. The issue will not be transparent and only
visible if the user examines the assisted-service container log.