WIP: data/bootstrap/files/usr/local/bin/bootkube: Drop cloud-cred operator render
We've had this since 24d8a48192 (add cloud-cred-operator as static
pod, 2019-10-04, #2447), originally to give Kuryr some scoped creds.
But I'm not clear on why that has to happend via 'render' and
cluster-bootstrap, instead of happening via the cluster-version
operator's usual resource reconciliation.  Dropping the render to see
how it plays in CI.