OpenStack: Fix IPv6 address configuration for bootstrap
When using an additional network in the cluster,
we end up having two conections of ethernet type, the
main one for the control plane and the additional one.
This would result in Network Manager not happy with the current
Network Manager profile. To avoid this issue, the IPv6
configuration for bootstrap is updated to match the
configuration currently applied on masters and workers[1]. This
config named `addr-gen-mode` ensures nodes get the same IPv6 and mac
addresses that were configured in the OpenStack Ports

[1] https://github.com/openshift/machine-config-operator/blob/release-4.15/templates/common/openstack/files/ipv6-config.yaml