pkg:types: add featuregates pkg
Adds a featuregate package to allow feature gating
based on openshift/api.

This commit copies code from other repos (acknowledged
in the headers) rather than importing because the sources
are well-contained and we want to minimize imports to
pkg/types as pkg/types is intended to be imported by other
codebases and by doing this we minimize their dependencies.

This code is mostly borrowed from:
https://github.com/openshift/cluster-config-operator/blob/636a2dc303037e2561a243ae1ab5c5b953ddad04/pkg/cmd/render/render.go#L153