aws byovpc: classifying wavelength subnets as edge
When installing a cluster in existing VPC, the subnet IDs must be
supplied, the zone attributes (type) for each one must be discovered,
and stored to be consumed by Machine set manifest assets generator.

This change force the Zone type wavelength-zone to be classified as
edge zone/subnets by installer, to create respective Machine Set
for that feature.

The subnet "internet facing discovery" requires to be updated to
classify the zone as "public" when the subnet is associated to an route
table which has a gateway resource id starting with 'cagw', following
the same logic of regular zones when classify as public when starts with
'igw'.