OCPBUGS-21777: baremetal: populate customDeploy in advance
Currently, we do not populate any deployment fields when creating
BareMetalHosts, relying on the Machine API to do it instead.
However, not setting customDeploy makes the BMH controller think that
we're doing an image-based deployment and configure its resources
accordingly. If MAPI is slow enough, the hosts become provisioned
before customDeploy is set, and the controller enters a loop trying and
failing to switch away from the image-based deployment.

TL;DR just set customDeploy on creation.