OpenStack: Fix dual-stack machines Spec to contain network
When the network ID or Name is not specified in the `controlPlanePort`
field, MAPO would create one Port per Subnet. In dual-stack clusters
we must have one unique dual-stack Port. This commit fixes the issue
by discovering the network ID from one of the subnets and adding it
to the Machine Spec.