images/installer-altinfra: initial Dockerfile
Creates the initial Dockerfile to create an image using the altinfra
build tag. The Dockerfie is necessary to have the Terraform-free
image built by ART.

The altinfra image is a temporary image to provide a Terraform-free
image for all providers that have implemented alternate infrastructure
providers. Once all providers have implemented non-Terraform solutions
it will not be necessary.