AGENT-337: Support both VIP and VIPs in AgentClusterInstall
The assisted-service is deprecating support for the singular APIVIP/IngressVIP
in https://github.com/openshift/assisted-service/pull/5501. When that merges
it will break the agent-installer since currently APIVIP is set when only
one VIP is configured. This change sets both API/Ingress VIP and VIPs and
will work before and after the assisted-service change merges.
Note that as a follow-on, https://github.com/openshift/installer/pull/7574
will remove the multiple settings from install-config override which are
not needed after the assisted-service change.