azure: add metadata info to resource group tags
When Hive runs the Installer destroy code, they don't always have all
the cluster metadata information the destroyer needs. Adding new fields
to the metadata is also a very painful process for them.

This PR adds some cluster metadata information to the cluster resource
group as tags, so that it can be discovered during destroy even if
azure-specific information is missing from the `metadata.json` file.

For now we are saving "region", "base domain resource group name", and
"network resource group name" but the mechanism will be useful if we
wish to add more data in the future without impacting Hive.