types/featuregates: don't panic if gate not in set
Updates the feature gate Enabled check to return false rather than
panic if a valid feature gate that is not part of a feature set is
not Enabled.

Instead of panicking, the function will return false. User-entered
feature gates are already validated in the install config. This panic
occurs when developers create feature gate checks in code, but the
feature gate is not part of a feature set, such as TechPreview. This
change will allow us to use the feature gates as part of the
CustomNoUpgrade feature set.