cluster metadata: add feature gate
Includes the feature gates for the cluster. These feature gates can
be used for testing bootstrap destroy functionality.

Including the feature gates and using them in this manner should not
create any issues. Particularly this should not affect hive, as long
as this metadata field is not used in the cluster destroy code (as
opposed to the bootstrap destroy code it is being used with here).

Regardless, it seems like the need for carrying featuregates in the
cluster metadata will be short lived while we move to new infrastructure
providers. Once these infrastructure providers are considered default,
it would be good to consider removing the feature gate in order to
keep cluster metadata lightweight.