azure: Enable storage account encryption
Added a field to accept the customer managed key vault id and
user assigned identity required for the encryption of storage
account.

This would require a few changes to the storage account as azure
requires[1] the account tier to be Premium and kind to be StorageV2
which is default and we do not set it.

[1] - https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/resources/storage_account#customer_managed_key