destroy/azure: fix dropped negation from error comparison
Somehow I missed the `!` when translating `err != context.Canceled` to
`!errors.Is(err, context.Canceled)`. However, it just so happens that
the comparison is not needed anymore since we're not purposefully
cancelling the context as a means of signifying the end for the callback
function.

Also changing the `immediate` parameter from `PollUntilContextCancel` to
`false`, so the callbacks are not run when the timeout has been reached
(otherwise they'll run at least once regardless).