OpenStack: UPI move resource names to inventory
OpenStack network related resource names which was defined in
common.ymal need to be moved to inventory.yaml, so the user will have a
chance to change them before running the network.yaml playbook. This is
needed as we will not have infraId to be added as part of the resource
names.