terraform: add build information to binaries
Store the hash of the go.mod in the built binary which we can then use
in hack/build.sh to determine if a provider needs to be rebuilt. Useful
when we import pre-built binaries from a container image.

Note: the -X flag allows setting a string variable within code, but we
are only using the value when inspecting the binary -- not within the
execution of code [1].

[1] https://github.com/golang/go/wiki/GcToolchainTricks#including-build-information-in-the-executable