network/manifest/aws: support user-defined cluster network MTU
The custom CNO manifest is automatically rendered when a edge compute
pool in AWS deployments. With this change, user-defined cluster network
MTU (field platform.aws.clusterNetworkMTU), this custom manifest must be
adapted allowing precedence when the field is added.

The validation flow raises a wanring when the values are out of generally
supported for aws edge zones (Local or Wavelength), and the manifest
is generated according to the user-defined in install-config.yaml,
otherwise the flow will keep as existing one: lowered the MTU when Local
or Wavelength zones are added, considering the overhead for each network
plugin.