ic/networking/mtu: validating field networking.clusterNetworkMTU
The field networking.clusterNetworkMTU is available initially only for
AWS deployment, allowing the customization in edge deployments.

Currently the installer automatically adjust the CNO manifest setting
the cluster network MTU for the entire cluster when the edge compute
pool (AWS Local or Wavelength zone) is added in install-config.yaml.

The field can allow users to customize the enforced value in certain
situations like deploying a cluster in newer AWS Local Zones which is
offering highest MTU when communicating with assets hosted in the
Region.

The user is responsible for those values, the warning is added when it
is changed in different situations.