OCPBUGS-23140, OCPBUGS-23305: Soften VIP validations for external load-balancer
Currently for cluster-managed load balancer we have a set of validations
ensuring that API and Ingress VIP are placed in a correct subnet as well
as they do not overlap.

For user-managed load balancer those validations are too strong as
customers may have external load balancers placed outside of networks
where cluster nodes live. Similarly, they may reuse the same IP for API
and Ingress VIP, and distribute traffic via other means (e.g. using the
port for the connection).

With this PR we are softening the validations if the configuration used
is indicating that external load balancer is used.

Closes: OCPBUGS-23140
Closes: OCPBUGS-23305