update azure cloudproviderconfig version
This commit updates the version of the azure cloud provider config.
We carry a copy of the struct in order to avoid vendoring unnecessary
dependencies. This copies in the current version of the config
from the openshift fork.

A new version is needed because the default for vmType has changed.

The new version also has renamed the CloudProviderRateLimit field,
so it has been removed from our manifest. The new value does not need
to be set because it defaults to False, which is desired.