openstack: Inject Bootstrap shim in CAPI manifests
This patch refactors how the bootstrap server's Ignition file is
uploaded to Glance when installing on OpenStack.

Before this patch, the `TerraformVariables` asset is responsible for
uploading the Ignition file to Glance and set Terraform to upload a
pointer to it.

However, CAPI manifest generation is or will be setting the full
Ignition file into the bootstrap Machine user data. CAPI manifest
generation happens in the `create manifests` target, in which no
modification to the target cluster can be done.

With this patch, the `TerraformVariables` step sets the full bootstrap
Ignition unchanged into tfvars.

With this patch, the upload of the Ignition file to Glance, and its
replacement with a shim in either tfvars or the CAPI manifests, happens
in the `Cluster` asset. The behaviour is coded into the `PreTerraform()`
and `PreCAPI()` hooks.