openstack [refactor]: Move VIP port tagging out of TFVars
Currently, Terraform variable generation happens whatever the install
flow. In other words: when installing with CAPI, the Installer will
trigger the generation of the Terraform variables anyway.

In preparation for a potential removal or deactivation of the
`TerraformVariables` step, this patch moves VIP-port-tagging to the
`PreTerraform()` and `PreCAPI()` hooks in the Cluster asset step.