pkg/infrastructure: add capi provider
Implements the infrastructure provider interface with the CAPI system.
This encapsulates the CAPI implementation similar to Terraform. It also
maintains pkg/infrastructure/platform.go (and build variants) as the
canonical source of truth for choosing an infrastructure provider.

This also adds an interface that cloud platforms utilizing the CAPI
provisioning should implement to provision additional resources.