infra/capi/aws: create int LB with CAPI
Switches the config so that CAPI creates the internal LB and the
external LB is created with direct SDK calls.

In the opposite configuration, one control plane would always be
NOTREADY. This seems preferred anyway because we will ALWAYS need
an internal LB, but an external LB is not needed with private clusters.

Ultimately this logic will be moved to the CAPA provider.