OCPBUGS-27453: baremetal: correct external_http_url for v6-only BMCs
Currently, in v4-primary dual-stack deployments, Ironic always uses its
v4 server URL as the image URL to pass to BMCs. If a BMC only has IPv6
networking, it is not going to work. This change checks the IP family
of the BMC and provides the correct external_http_url.

Differences in the backport:
* getParsedURL copied from baremetal-operator to avoid catching up with
  a long history of changes in baremetal-operator before this function
  became public.

(cherry picked from commit 99c460281991b7814fa412735fee406571211316)