vsphere-fix convert if only provided name
In 4.12 and prior versions of the installer
allowed non-pathed names of vCenter objects.
In 4.13+ that is no longer allowed and absolute full
paths are required.
This change is causing issues with ACM/Hive. The
only resolution (not ideal) is to connect
to vCenter and determine from potentially only the
objects name to determine the absolute full path.

If there is no available connection to vCenter in the
case of AI or Agent installer this fails back to
only merging paths.

Move existing functions from private to public
to assist hive - reducing copy and paste.