OCPBUGS-27156: GCP Destroy cleanup only records/zones created by the installer
** The cluster domain was added to the metadata previously to find private zones that were
not created by the installer. The records were added to a private zone in this domain. The
problem occurs when clusters with the same name are used for installation. The infra ID is
not used during the zone creation or zone lookup. The same zone (with the same domain) is
assumed to be the one the installer (destroyer) is looking for. The fix includes skipping the
addition of the domain to the metadata so that it is not used for comparison on destroy.