OCPBUGS-27737: aws-edge-zones: preventing errors when discovering phase
AWS Wavelength Zones have been introduced as part of edge-zones in 4.15.
The function used to discover the valid edge zones in the region is
prematurely returning errors when the region haven't available any of
Local Zones or Wavelength Zones, which could cause issues in regions
which support only one edge zone type.

https://issues.redhat.com//browse/OCPBUGS-27737