OpenStack: fix controlPlanePort validation
When the controlPlanePort has no network defined, we attempt to
validate the network using the empty name and id, this results
in multiple networks found, which breaks the pre-flight validation
resulting in it being skipped. This commit fixed the issue, by only
running validation on that field if it has content defined.