OpenStack: Fix dualstack with external load-balancer
When the cluster is created with an external load-balancer,
no api or ingress VIP Ports exists, but for dual-stack
installations we are expecting those Ports to pre-exists,
resulting in failure to add security groups to those Ports.
This commit fixes the issue by ensuring to only attach the
security group when no external lb is configured and dual-stack
is used.