terraform vs infra sdk capv folder create
With the terraform vsphere provider its nearly
impossible to create multiple nested folders.

The introduction using govmomi to create the
pre-requisite objects in vCenter elivates
that problem.

To resolve the validation issue the folder
function is split between two files.

If terraform is used the confirmation that the folder
exist is still present.

If capv/altinfra is used the folder does not need
to be present and continues if not.

These files should only be needed until terraform
is removed entirely.