Use top-level context for create commands
Adds a top-level context which is used when generating assets
and is passed to PostRun functions for waiting for install complete.

The context's cancel function is passed to the logrus exit handler
to ensure that the context is cancelled at any exit point.

The main motivation for introducing this context is to ensure a
clean shutdown of any cluster-api controllers.