OCPBUGS-29469: fix Azure API SKU calls timing out - part 2
I erroneously assumed that Azure was not returning context errors but
that was not actually the case. It was happening as a result of the way
we structured the API call:

```
for page, err := client.List(ctx); page.NotDone(); err =
page.NextWithContext(ctx) {
    if err != nil {
	return nil, fmt.Errorf("failed to list SKUs: %w", err)
    }
}
```

If the `client.List` call fails (e.g, context deadline, auth error),
`page.NotDone()` will return `false` and we'll never enter the `for`
loop to check for `err`. Instead, we should make the `List` call and
check for error *before* entering the loop.