Fix CAPI bootstrap deletion
The tear down of local capi controller is happening
right after triggering the deletion of the
bootstrap machine, leaving almost no time
for the machine deletion event to be handled
resulting in a leftover bootstrap machine.
This commit fixes the issue by checking up to
2 minutes if the bootstrap machine was deleted
prior to destroy the local control plane.