CORS-3295: hack: add script to verify capi CRDs
This script will be used to create a test in CI to check that the capi
CRDs in data/data/cluster-api match what's expected from the provider
version used in cluster-api/providers.

As a side effect, this script can be used to generate the CRD, allowing
a dev to then commit the result.