azure: use filter when listing SKUs
Prior to this commit, most calls to list() in the azure resource
sku client were using a package version that did not support
filtering list calls. This commit updates all calls to use the
location filter and greatly speeds up the performance for validation.
In local testing, the total time to generate ignition configs was almost
2 minutes prior to these changes, which reduced the running time to
20 seconds.