CORS-3142: capi: write manifests to disk during create manifests
Write CAPI manifests to disk during create manifests so that they can be
user edited and users can also provide their own set of manifests. In
general, we think of manifests as an escape hatch that should be used
when a feature is missing from the install config, and users accept the
degraded user experience of editing manifests in order to achieve
non-install-config-supported functionality.