cmd: plumb in signal handler
Plumbs in the controller runtime signal handler to trap user interrupts
and gracefully shutdown. The signal handler can only be registered
once in a program, so it is added to main so that it can be utilized
throughout the codebase.

The signal handler provides a context that is cancelled upon interrupt,
and can therefore be used to trigger a graceful shutdown.

This commit plumbs the signal handler context through to all commands,
even if there is no plan to use it, for consistency and potential
future convenience.