Add context and signal handler to all commands
Adds a context and signal handler to all commands. By doing this
we setup the future possibility for commands to utilize a top-level
context (if needed) and, furthermore, escape from the interrupt trap
set in the main function, so that users do not need to press CTRL+C
twice to interrupt the program.