feat/elbv2/tg: CAPA changes to set cluster TG health check
CAPA sets the listener protocol and port as the target health check
configuration (protocol and port) for all listeners. When the
HealthCheckProtocol is set to the LB, only the API's target group will be
configured to use the desired protocol.

Motivated by keeping the compatibility with Terraform, the upstream
PR#4849 has been created to implement the customization for API target
group, and more flexibility for the additional listeners.

This change sets the health check for all listeners used by Load Balancers
managed by CAPA, both primary and secondary,
as well the additional listeners for MCS.