HIVE-2476: OpenStack.MachineSets(): Allow nil Replicas
A nil pointer exception occurs when the OpenStack MachineSets() func is
called with `Replicas==nil`, as it would be when hive is attempting to
generate autoscaling MachineSets -- we set the replica count later on
after we find out how many MachineSets to spread the autoscaling min/max
across.

Add a safeguard to default to zero in this scenario.

Needed by HIVE-2476