Use new build tag to differentiate dynamic build
When we linked to libvirt directly, having the libvirt platform in the
build was synonymous with a dynamically linked build. Now that we use a
pure-Go libvirt library, the build type can be independent of whether
libvirt is included.

Once the baremetal installer build no longer includes the libvirt
platform, we still want it to be dynamically linked.