OCPBUGS-27965: escape '%' in proxy settings
The 'DefaultEnvironment' config from systemd accepts some % specifiers
for expansion [1]. Because of that, proxy configurations that make use
of that character need to be escaped with another '%'. However, during
proxy validation, the Installer parses the URI and fails when it
encounters a "%%" with an `invalid URL escape` error.

To mitigate the problem, we introduce a new `replace` function to the
templating system and use it to substitute "%" with "%%" causing the
following config:

```
proxy:
    httpProxy: http://user%40test:redhat@10.0.0.1:3128
    httpsProxy: http://user%40test:redhat@10.0.0.1:3128
```

to become the `10-default-env.conf`:

```
[Manager]
DefaultEnvironment=HTTP_PROXY="http://user%%40test:redhat@10.0.0.1:3128"
DefaultEnvironment=HTTPS_PROXY="http://user%%40test:redhat@10.0.0.1:3128"
```

[1] https://man7.org/linux/man-pages/man5/systemd-system.conf.5.html