Don't template systemd dependencies
In systemd unit files, After= only affects ordering and can reference
units that don't exist or exist but are not enabled, so there is no
reason to use templating.

In the appliance use case we will need the agent image that is installed
to disk to contain systemd units that work for either installing or
adding a node (when the appropriate config is added), so both units must
have the correct ordering.