Configure install/add-node dependencies in systemd
There are two pairs of services that are always needed together in the
agent ISO. For installing a cluster, agent-register-cluster.service
followed (later) by start-cluster-installation.service. For adding a
node to a cluster, agent-import-cluster.service followed (later) by
agent-add-node.service.

Reflect these dependencies in the systemd units' Install sections, so
that we only need to enable either start-cluster-installation.service or
agent-add-node.service to ensure all of the required services are
enabled. This will simplify the implementation of adding a node via the
appliance, where one flow or the other will need to be triggered in
response to the config ISO being attached. Do not make either unit a
requirement of multi-user.target, as they conflict. That allows us to
enable both units (i.e. execute their Install sections) in the ignition.

This ensures that when we start start-cluster-installation.service upon
seeing a config ISO attached to the appliance,
agent-register-cluster.service also gets started. This service was
previously inadvertantly disabled by
9716c1f6046e225647e67af616a211f47b995e90.

Up to then, we also relied on enabling
start-cluster-installation.service in the unconfigured ignition.
However, due to the remnants of an implementation that existed prior to
204c4ae561127220d296a52a575d2103ff18f90a, there is still code in
load-config-iso.sh to enable the service after the config drive is
attached. This would be needed for an interactive ISO, but for the
applicance we changed to enabling the service in the ignition.

In future we will need to choose whether to start
start-cluster-installation.service or agent-add-node.service based on
the contents of the config drive, so continue to do this at runtime
rather than simply re-enabling the former in the unconfigured ignition.