capi: add timeout when provisioning
Adds a 15m timeout to the infrastructure provisioning and machine
provisioning stages of CAPI, so that the controllers do not
spin indefinitely in the case of a failure. 15m is an arbitrary
value, but the criteria for the timeout should be based on the
balance of ample time to provision the resources with not
making users wait too long if something goes wrong.